requirements.txt
duckdb>=1.1.0
neo4j>=5.24.0
python-dotenv>=1.0.1
apache-airflow>=2.10.0